src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aniso8601=8.0.0=py_0
appdirs=1.4.4=py_0
appnope=0.1.0=py38_1001
attrs=20.1.0=py_0
backcall=0.2.0=py_0
black=19.10b0=py_0
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py38h32f6830_0
cffi=1.14.2=py38hed5b41f_0
click=7.1.2=py_0
cryptography=3.1=py38h52adbb4_0
decorator=4.4.2=py_0
flask=1.1.2=py_0
flask-jwt-extended=3.24.1=pyh9f0ad1d_0
flask-restful=0.3.8=py_0
idna=2.10=pyh9f0ad1d_0
ipython=7.18.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
itsdangerous=1.1.0=py_0
jedi=0.17.2=py38_0
jinja2=2.11.2=py_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
markupsafe=1.1.1=py38h1de35cc_1
mypy_extensions=0.4.3=py38_0
ncurses=6.2=h0a44026_1
openssl=1.1.1g=haf1e3a3_1
parso=0.7.0=py_0
pathspec=0.7.0=py_0
pexpect=4.8.0=py38_1
pickleshare=0.7.5=py38_1001
pip=20.2.2=py38_0
prompt-toolkit=3.0.7=py_0
ptyprocess=0.6.0=py38_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.6.1=py_0
pyjwt=1.7.1=py_0
python=3.8.5=h26836e1_1
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytz=2020.1=pyh9f0ad1d_0
readline=8.0=h1de35cc_0
regex=2020.7.14=py38haf1e3a3_0
setuptools=49.6.0=py38_0
six=1.15.0=py_0
sqlite=3.33.0=hffcf06c_0
tk=8.6.10=hb0a8c7a_0
toml=0.10.1=py_0
traitlets=4.3.3=py38_0
typed-ast=1.4.1=py38h1de35cc_0
typing_extensions=3.7.4.3=py_0
wcwidth=0.2.5=py_0
werkzeug=1.0.1=py_0
wheel=0.35.1=py_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
